Attempted leaked semaphore fix for macos
